Metodo de busca ordenada de testes / ajuste na atualização de sequencia

+++ back/rail/src/main/java/com/oficinabr/rail/dto/TestDTO.java
@@ -2,18 +2,57 @@ package com.oficinabr.rail.dto;
 
 import com.oficinabr.rail.entity.Test;
 
-public record TestDTO(String id, Integer injectorNumber, Long sequence, String comments, String date, Float resistance, Float inductance, Float isolation,
-		Float halfLoad, Float fullLoad, Float idling, Float preInjection, Float halfLoadReturn, Float fullLoadReturn,
-		Float idlingReturn, Float preInjectionReturn, String planId, String injectorId, String injectorModel,
-		String vehicleId, String vehiclePlate, String imaCode, String serviceOrder
+public record TestDTO(
+		String id, 
+		Integer injectorNumber, 
+		Integer sequence, 
+		String comments, 
+		String date, 
+		Float resistance, 
+		Float inductance, 
+		Float isolation,
+		Float halfLoad,
+		Float fullLoad, 
+		Float idling, 
+		Float preInjection, 
+		Float halfLoadReturn, 
+		Float fullLoadReturn,
+		Float idlingReturn, 
+		Float preInjectionReturn, 
+		String planId, 
+		String injectorId, 
+		String injectorModel,
+		String vehicleId, 
+		String vehiclePlate, 
+		String imaCode, 
+		String serviceOrder
 
 ) {
 	public TestDTO(Test t) {
-		this(t.getId(), t.getInjectorNumber(), t.getSequence(), t.getComments(), t.getDate(), t.getResistance(), t.getInductance(), t.getIsolation(),
-				t.getHalfLoad(), t.getFullLoad(), t.getIdling(), t.getPreInjection(), t.getHalfLoadReturn(),
-				t.getFullLoadReturn(), t.getIdlingReturn(), t.getPreInjectionReturn(), t.getPlan().getId(),
-				t.getInjector().getId(), t.getInjector().getModel(), t.getVehicle().getId(), t.getVehicle().getPlate(),
-				t.getImaCode(), t.getServiceOrder());
+		this(
+		t.getId(), 
+		t.getInjectorNumber(), 
+		t.getSequence(), 
+		t.getComments(), 
+		t.getDate(), 
+		t.getResistance(), 
+		t.getInductance(), 
+		t.getIsolation(),
+		t.getHalfLoad(), 
+		t.getFullLoad(), 
+		t.getIdling(), 
+		t.getPreInjection(), 
+		t.getHalfLoadReturn(),
+		t.getFullLoadReturn(), 
+		t.getIdlingReturn(), 
+		t.getPreInjectionReturn(), 
+		t.getPlan().getId(),
+		t.getInjector().getId(), 
+		t.getInjector().getModel(), 
+		t.getVehicle().getId(), 
+		t.getVehicle().getPlate(),
+		t.getImaCode(), 
+		t.getServiceOrder());
 	}
 
 }
